chore: tidied up the code base by getting rid of GameManager.cs and GameDatabase.cs
