+++ Assets/ARC_CityBuilder/Materials/Script/CustomBuildings/TestShelterOrderUI.cs
@@ -4,7 +4,7 @@ public class TestShelterOrderUI : MonoBehaviour
 {
     public void OnAdvanceDayClicked()
     {
-        GameDatabase.Instance.AdvanceDay();
+        BuildingSystem.Instance.AdvanceDay();
         Debug.Log("[ShelterOrderUI] Advance Day clicked — food orders generated.");
     }
 }
