+++ Assets/ARC_CityBuilder/Materials/Script/Master/BuildingSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2a5c8dae237ac554abde649bac21d930
+guid: 32598ec39d23ea641866b9520703ad0b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
