+++ Assets/ARC_CityBuilder/Materials/Script/Master/DisasterManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dc6bb17f4423da047a9b159b14086e87
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
