+++ Assets/ARC_CityBuilder/Materials/Script/Master/MasterGameManager.cs
@@ -291,188 +291,7 @@ public class MasterGameManager : DefaultGameManager
     }
 }
 
-/// <summary>
-/// Manages round progression and timing
-/// </summary>
-public class RoundManager : MonoBehaviour
-{
-    [Header("Round Settings")]
-    public float startRoundDelay = 1f;
-    public float endRoundDelay = 1f;
-    
-    /// <summary>
-    /// Handle start of round behaviors
-    /// </summary>
-    public IEnumerator StartRound(int roundNumber, int dayNumber)
-    {
-        Debug.Log($"Round {roundNumber} (Day {dayNumber}) begins!");
-        
-        // Update UI
-        var uiManager = FindObjectOfType<UIManager>();
-        if (uiManager != null)
-        {
-            uiManager.UpdateRoundText(roundNumber, dayNumber);
-        }
-        
-        // Add any other start of round logic here
-        
-        // Pause for transitions
-        yield return new WaitForSeconds(startRoundDelay);
-    }
-    
-    /// <summary>
-    /// Handle end of round behaviors
-    /// </summary>
-    public IEnumerator EndRound()
-    {
-        Debug.Log("Round ends!");
-        
-        // Add any end of round logic here
-        
-        // Pause for transitions
-        yield return new WaitForSeconds(endRoundDelay);
-    }
-}
 
-/// <summary>
-/// Manages all disaster systems
-/// </summary>
-public class DisasterManager : MonoBehaviour
-{
-    [Header("Disaster Systems")]
-    public FloodManager floodManager;
-    
-    [Header("Settings")]
-    public float disasterPhaseDelay = 2f;
-    
-    private void Awake()
-    {
-        // Get required components
-        if (floodManager == null)
-            floodManager = FindObjectOfType<FloodManager>();
-    }
-    
-    /// <summary>
-    /// Process all disaster events for the current round
-    /// </summary>
-    public IEnumerator ProcessDisasters()
-    {
-        Debug.Log("Processing disaster events...");
-        
-        // Process flooding
-        if (floodManager != null)
-        {
-            floodManager.SimulateFlooding();
-        }
-        
-        // Other disasters can be added here
-        
-        yield return new WaitForSeconds(disasterPhaseDelay);
-    }
-}
 
-/// <summary>
-/// Manages building-related systems and communication
-/// </summary>
-public class BuildingSystem : MonoBehaviour
-{
-    // References to required systems
-    private IBuildingManager _buildingManager;
-    private List<Building> _communities = new List<Building>();
-    private List<Building> _shelters = new List<Building>();
-    private List<Building> _kitchens = new List<Building>();
-    private List<Building> _otherBuildings = new List<Building>();
-    
-    private void Start()
-    {
-        _buildingManager = Dependencies.GetOptional<IBuildingManager>();
-        
-        // Import existing buildings from GameDatabase if it exists
-        var gameDatabase = FindObjectOfType<GameDatabase>();
-        if (gameDatabase != null)
-        {
-            _communities.AddRange(gameDatabase.GetAllCommunities());
-            _shelters.AddRange(gameDatabase.GetAllShelters());
-            _kitchens.AddRange(gameDatabase.GetAllKitchens());
-            _otherBuildings.AddRange(gameDatabase.GetAllGenericBuildings());
-        }
-    }
-    
-    /// <summary>
-    /// Register a new building in the system
-    /// </summary>
-    public void RegisterBuilding(Building building, string type)
-    {
-        switch (type.ToLower())
-        {
-            case "community":
-                if (!_communities.Contains(building))
-                    _communities.Add(building);
-                break;
-            case "shelter":
-                if (!_shelters.Contains(building))
-                    _shelters.Add(building);
-                break;
-            case "kitchen":
-                if (!_kitchens.Contains(building))
-                    _kitchens.Add(building);
-                break;
-            default:
-                if (!_otherBuildings.Contains(building))
-                    _otherBuildings.Add(building);
-                break;
-        }
-    }
-    
-    /// <summary>
-    /// Update all buildings after disaster events
-    /// </summary>
-    public void UpdateAllBuildings()
-    {
-        // Notify communities of flood changes
-        foreach (var community in _communities)
-        {
-            if (community != null && community.TryGetComponent<CommunityLogic>(out var logic))
-            {
-                logic.CheckFloodStatusAndUpdateOrder();
-            }
-        }
-        
-        // Reset food orders in shelters
-        foreach (var shelter in _shelters)
-        {
-            if (shelter != null && shelter.TryGetComponent<ShelterLogic>(out var logic))
-            {
-                logic.ClearFoodStorage();
-                logic.GenerateFoodOrderDebug();
-            }
-        }
-        
-        // Notify kitchens of new orders
-        foreach (var kitchen in _kitchens)
-        {
-            if (kitchen != null && kitchen.TryGetComponent<ProductionWalkerComponent>(out var productionWalker))
-            {
-                productionWalker.TryRestartDelivery();
-            }
-        }
-    }
-    
-    /// <summary>
-    /// Get all buildings of a specific type
-    /// </summary>
-    public IReadOnlyList<Building> GetBuildingsByType(string type)
-    {
-        switch (type.ToLower())
-        {
-            case "community":
-                return _communities;
-            case "shelter":
-                return _shelters;
-            case "kitchen":
-                return _kitchens;
-            default:
-                return _otherBuildings;
-        }
-    }
-}
+
+
